requirements.txt
opencv-python
groq
pillow
python-dotenv
requests
streamlit
streamlit-option-menu